Nasty hack to fix routing in go CC
The combined "/v3/buildpacks AND method GET" ACLs don't work. The only
way to combine ACLs is in the use_backend block which we don't have
control over in the haproxy-boshrelease. For now (until
https://github.com/cloudfoundry/haproxy-boshrelease/pull/271 is merged),
we can use the frontend_config block to create our own acls and
use_backend blocks that call backends created in the normal way. This
relies on some slightly brittle shasums (the backends are named based on
the key from the routed_backend_servers property) but it should work for
now