Use bash variable defaulting to test for empty
We use set -u at the top of this script which exits on unbound
variables. This gets triggered even if that variable is only being used
for existance tests (ie. if [ -z "$VAR" ]). By using :- we can set the
value to "" if the variable is not set which gets correctly picked up by
-z and we get the logic we wanted