Adopt the performance tests to run on the CC VM
This removes the routing layer, concourse and public internet from error
sources for unreliable results as we run the tests on the CC vm against
the local nginx. We also run against plain http so that we do not
measure https handshakes. Tests now also run on a non burstable instance
so that cpu performance stays the same.