Improve retry mechanism to upload state on failure
In case of an error the attemts parameter leads to a restart of the bbl
opereratio without persisting state.
The only way to circumvent this is to run the task multiple times as
concourse has no property to handle this.
Since nested on_failure hooks are also not allowed we have to run the
idempotent tasks multiple times and upload the state after every try.